+++ src/main/kotlin/dev/s7a/ktinventory/AbstractKtInventoryFetched.kt
@@ -55,6 +55,14 @@ abstract class AbstractKtInventoryFetched<T : AbstractKtInventoryFetched<T, C>, 
         page: Page<C, KtInventoryButton<Entry<T, C>>>,
     ): Entry<T, C>
 
+    final override fun button(
+        slot: Int,
+        item: KtInventoryButton<KtInventoryBase>,
+    ) {
+        require(slot !in paginates) { "button slot must not be used as a pagination slot (actual: $slot)" }
+        super.button(slot, item)
+    }
+
     /**
      * Sets slots to be used for pagination.
      *
@@ -72,6 +80,8 @@ abstract class AbstractKtInventoryFetched<T : AbstractKtInventoryFetched<T, C>, 
      * @since 2.2.0
      */
     fun paginateSlot(slots: Iterable<Int>) {
+        val buttons = this.buttons
+        require(slots.none { it in buttons }) { "pagination slots must not contain fixed button slots" }
         this.paginates.addAll(slots)
     }